Divided the requestsTests.cpp in 3 files : getTests.cpp, deleteTests.cpp and putTests.cpp
